Task 11 Done!

+++ restaurants/views.py
@@ -59,6 +59,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect("signin")
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -73,8 +75,11 @@ def restaurant_create(request):
     return render(request, 'create.html', context)
 
 def item_create(request, restaurant_id):
+
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not request.user.is_staff and request.user != restaurant.owner:
+        return redirect("unauthorized")
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -89,8 +94,11 @@ def item_create(request, restaurant_id):
     return render(request, 'item_create.html', context)
 
 def restaurant_update(request, restaurant_id):
+
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     form = RestaurantForm(instance=restaurant_obj)
+    if not request.user.is_staff and request.user != restaurant_obj.owner:
+        return redirect("unauthorized")
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
         if form.is_valid():
@@ -103,6 +111,12 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not request.user.is_staff:
+        return redirect("unauthorized")
     restaurant_obj.delete()
     return redirect('restaurant-list')
+
+def unauthorized(request):
+    return render(request,"unauthorized.html")
